update clerk api spec

+++ src/models/stubs_period_verification_period_admin.rs
@@ -23,14 +23,14 @@ pub struct StubsPeriodVerificationPeriodAdmin {
         with = "::serde_with::rust::double_option",
         skip_serializing_if = "Option::is_none"
     )]
-    pub attempts: Option<Option<i64>>,
+    pub attempts: Option<Option<i32>>,
     #[serde(
         rename = "expire_at",
         default,
         with = "::serde_with::rust::double_option",
         skip_serializing_if = "Option::is_none"
     )]
-    pub expire_at: Option<Option<i64>>,
+    pub expire_at: Option<Option<i32>>,
 }
 
 impl StubsPeriodVerificationPeriodAdmin {
@@ -43,7 +43,7 @@ impl StubsPeriodVerificationPeriodAdmin {
         }
     }
 }
-///
+
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
 pub enum Status {
     #[serde(rename = "verified")]
@@ -55,7 +55,7 @@ impl Default for Status {
         Self::Verified
     }
 }
-///
+
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
 pub enum Strategy {
     #[serde(rename = "admin")]
