+++ src/models/stubs_period_verification_period_link.rs
@@ -23,9 +23,9 @@ pub struct StubsPeriodVerificationPeriodLink {
         with = "::serde_with::rust::double_option",
         skip_serializing_if = "Option::is_none"
     )]
-    pub attempts: Option<Option<i64>>,
+    pub attempts: Option<Option<i32>>,
     #[serde(rename = "expire_at")]
-    pub expire_at: i64,
+    pub expire_at: i32,
     #[serde(rename = "verified_at_client", skip_serializing_if = "Option::is_none")]
     pub verified_at_client: Option<String>,
 }
@@ -34,7 +34,7 @@ impl StubsPeriodVerificationPeriodLink {
     pub fn new(
         status: Status,
         strategy: Strategy,
-        expire_at: i64,
+        expire_at: i32,
     ) -> StubsPeriodVerificationPeriodLink {
         StubsPeriodVerificationPeriodLink {
             status,
@@ -45,7 +45,7 @@ impl StubsPeriodVerificationPeriodLink {
         }
     }
 }
-///
+
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
 pub enum Status {
     #[serde(rename = "unverified")]
@@ -63,7 +63,7 @@ impl Default for Status {
         Self::Unverified
     }
 }
-///
+
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
 pub enum Strategy {
     #[serde(rename = "email_link")]
