+++ src/models/stubs_period_verification_period_totp.rs
@@ -23,14 +23,14 @@ pub struct StubsPeriodVerificationPeriodTotp {
         with = "::serde_with::rust::double_option",
         skip_serializing_if = "Option::is_none"
     )]
-    pub attempts: Option<Option<i64>>,
+    pub attempts: Option<Option<i32>>,
     #[serde(
         rename = "expire_at",
         default,
         with = "::serde_with::rust::double_option",
         skip_serializing_if = "Option::is_none"
     )]
-    pub expire_at: Option<Option<i64>>,
+    pub expire_at: Option<Option<i32>>,
 }
 
 impl StubsPeriodVerificationPeriodTotp {
@@ -43,7 +43,7 @@ impl StubsPeriodVerificationPeriodTotp {
         }
     }
 }
-///
+
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
 pub enum Status {
     #[serde(rename = "unverified")]
@@ -57,7 +57,7 @@ impl Default for Status {
         Self::Unverified
     }
 }
-///
+
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
 pub enum Strategy {
     #[serde(rename = "totp")]
